+++ app/(team)/[id]/todo/_components/todo.tsx
@@ -8,15 +8,15 @@ import { usePathname } from "next/navigation";
 import { useState } from "react";
 import useOverlay from "@/app/_hooks/useOverlay";
 import SideBarWrapper from "@/app/_components/sidebar";
-import TodoDetail from "@/app/(team)/[id]/todo/todoDetail";
 import { convertIsoToDateToKorean } from "@/app/_utils/IsoToFriendlyDate";
 import Image from "next/image";
 import tasksKey from "@/app/(team)/[id]/todo/_components/api/queryFactory";
 import Popover from "@/app/_components/Popover";
-import AddTaskModal from "@/app/(team)/[id]/todo/AddTask";
 import TodoItem from "@/app/(team)/[id]/todo/_components/TodoItem";
 import { useGetGroupList, useGetTodoItems } from "@/app/(team)/[id]/todo/_components/api/useQuery";
 import { useTodoCheckMutation } from "@/app/(team)/[id]/todo/_components/api/useMutation";
+import AddTaskModal from "@/app/(team)/[id]/todo/_components/AddTask";
+import TodoDetail from "@/app/(team)/[id]/todo/_components/todoDetail";
 
 type ClientTodoProps = {
 	groupId: number;
